Move Search*Mocks to mocks.dart
