+++ app/test/frontend/mocks.dart
@@ -3,11 +3,14 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:pub_dartlang_org/dartdoc/models.dart';
+import 'package:pub_dartlang_org/frontend/search_service.dart';
 import 'package:pub_dartlang_org/scorecard/backend.dart';
 import 'package:pub_dartlang_org/scorecard/models.dart';
 import 'package:pub_dartlang_org/shared/analyzer_service.dart';
 import 'package:pub_dartlang_org/shared/analyzer_client.dart';
 import 'package:pub_dartlang_org/shared/dartdoc_client.dart';
+import 'package:pub_dartlang_org/shared/search_client.dart';
+import 'package:pub_dartlang_org/shared/search_service.dart';
 
 class ScoreCardBackendMock implements ScoreCardBackend {
   @override
@@ -101,4 +104,37 @@ class DartdocClientMock implements DartdocClient {
       {Duration timeout}) async {
     return null;
   }
+}
+
+class SearchClientMock implements SearchClient {
+  final Function searchFun;
+  SearchClientMock({this.searchFun});
+
+  @override
+  Future<PackageSearchResult> search(SearchQuery query) async {
+    if (searchFun == null) {
+      throw new Exception('no searchFun');
+    }
+    return (await searchFun(query)) as PackageSearchResult;
+  }
+
+  @override
+  Future triggerReindex(String package, String version) async {}
+
+  @override
+  Future close() async {}
+}
+
+class SearchServiceMock implements SearchService {
+  final Function searchFun;
+
+  SearchServiceMock(this.searchFun);
+
+  @override
+  Future<SearchResultPage> search(SearchQuery query) async {
+    return (await searchFun(query)) as SearchResultPage;
+  }
+
+  @override
+  Future close() async => null;
 }
